Fixed bug with Helm HUD power display not updating correctly

+++ MoreCyclopsUpgrades/Patchers/CyclopsUpgradeConsoleHUDManager_RefreshScreen_Patcher.cs
@@ -14,7 +14,7 @@
             if (upgradeConsole == null)
                 return; // safety check
 
-            float currentReservePower = PowerCharging.GetTotalReservePower(upgradeConsole.modules);
+            int currentReservePower = PowerCharging.GetTotalReservePower(upgradeConsole.modules);
 
             float currentBatteryPower = __instance.subRoot.powerRelay.GetPower();
